+++ lib/src/in_app_webview.dart
@@ -162,7 +162,7 @@ class InAppWebView extends StatefulWidget implements WebView {
 
   @override
   final Future<bool> Function(InAppWebViewController controller,
-      CreateWindowRequest onCreateWindowRequest) onCreateWindow;
+      CreateWindowRequest createWindowRequest) onCreateWindow;
 
   @override
   final void Function(InAppWebViewController controller) onCloseWindow;